refactor: use same r2dbc apis for LocationEtlRepository used in repository module and use awaitOneOrNull instead of awaitSingleOrNull for query with one possible result

+++ etl/src/main/kotlin/com/github/ben_lc/sapin/etl/EtlProperties.kt
@@ -32,7 +32,7 @@ data class GeopackageProperties(val location: List<LocationGeopackageProperties>
 data class LocationGeopackageProperties(
     val tableName: String,
     val nameColumnName: String,
-    val level: Int,
+    val level: Short,
     val srcIdColumnName: String,
     val isoIdColumnName: String? = null,
     val levelNameColumnName: String? = null,
